requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
av==15.1.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.0
ctranslate2==4.4.0
cycler==0.12.1
docopt==0.6.2
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.117.1
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
httplib2==0.31.0
httptools==0.6.4
huggingface-hub==0.35.1
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
julius==0.2.7
kiwisolver==1.4.7
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.43.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
onnxruntime==1.19.2
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
optuna==4.5.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
primePy==1.3
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
semver==3.0.4
sentence-transformers==5.1.1
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
tomli==2.2.1
torch==2.8.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.56.2
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
whisperx @ git+https://github.com/m-bain/whisperx.git@2d9ce44329ae73af2520196d31cd14b6192ace44
yarl==1.20.1
zipp==3.23.0
